pyproject.toml
[project]
name = "prep-buddy-ai"
version = "1.0.0"
description = "Prep Buddy AI — a document-based learning assistant built with Claude, LangGraph, ChromaDB, and Streamlit"
requires-python = ">=3.11, <3.13"

dependencies = [
    # Anthropic
    "anthropic>=0.40.0",

    # LangChain
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",

    # LangGraph
    "langgraph>=1.0.0",

    # Vector store
    "chromadb>=0.5.0",

    # Embeddings
    "sentence-transformers>=3.0.0",

    # Document loaders
    "pypdf>=4.0.0",
    "unstructured>=0.15.0",
    "python-docx>=1.0.0",

    # UI
    "streamlit>=1.38.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
]
